separate button from language to a different component
