+++ src/components/Header/headStyles.ts
@@ -2,17 +2,6 @@ const className = {
   header: 'flex lg:flex-row lg:justify-evenly lg:items-center',
   mobileHeader: 'flex flex-col pt-5 items-center',
   mobileLogoButton: 'flex  lg:hidden items-center',
-  containerMobileLanguageButton:
-    'ontainerMobilelanguageButton flex relative pl-5',
-  mobilelanguageButton:
-    'w-[60px] h-[45px] bg-primary-base hover:cursor-pointer rounded-[100px]  relative z-10  flex items-center justify-around',
-  mobileLanguagetext: 'text-secondary-darker font-semibold text-xl pb-0.5',
-  containerLanguageButtons: 'flex relative pl-5 ',
-  languageButton:
-    'flex w-[134px] h-[45px]  hover:cursor-pointer bg-primary-base items-center justify-around rounded-[100px] relative z-10',
-  mobileMenuHeder: 'absolute z-0 ',
-  menuHeader: 'flex absolute z-0 ',
-  languageText: 'text-secondary-darker font-semibold font-jost',
   logoImage: 'logoImage sm:hidden',
   logoImageTextSm: 'logoImageTextSm hidden sm:block',
   logoImageTextLg: 'logoImageTextLg hidden lg:block',
