--- src/components/LangButtons/langButtons.types.ts
@@ -1,9 +1,0 @@
-export interface TypeLanguage {
-  [key: string]: {
-    [key: string]: string
-  }
-}
-
-export interface LanguageButtonProps {
-  isMobile?: boolean
-}
