+++ src/components/LanguagesButton/languagesButton.types.ts
@@ -1,0 +1,3 @@
+export interface LanguagesButtonsProps {
+  isMobile?: boolean
+}
