+++ src/components/LanguagesButton/languagesButtonStyles.ts
@@ -1,0 +1,14 @@
+const className = {
+  containerMobileLanguageButton:
+    'ontainerMobilelanguageButton flex relative pl-5',
+  mobilelanguageButton:
+    'w-[60px] h-[45px] bg-primary-base hover:cursor-pointer rounded-[100px]  relative z-10  flex items-center justify-around',
+  mobileLanguagetext: 'text-secondary-darker font-semibold text-xl pb-0.5',
+  mobileMenuHeder: 'absolute z-0 ',
+  containerLanguageButtons: 'flex relative pl-5 ',
+  languageButton:
+    'flex w-[134px] h-[45px]  hover:cursor-pointer bg-primary-base items-center justify-around rounded-[100px] relative z-10',
+  languageText: 'text-secondary-darker font-semibold font-jost',
+  menuHeader: 'flex absolute z-0 ',
+}
+export default className
