+++ src/components/LanguagesMenu/languagesMenu.types.ts
@@ -1,0 +1,3 @@
+export interface LangMenuProps {
+  isMobile?: boolean
+}
